emotion_chat/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
python-dotenv>=1.0
psycopg2-binary>=2.9
pgvector>=0.3

daphne>=4.0
channels>=4.0
channels-redis>=4.1
redis>=5.0
django-redis>=5.4
celery[redis]>=5.3

langgraph>=0.2.50
langchain>=0.3
langchain-core>=0.3
langchain-google-genai>=2.0
langchain-openai>=0.2

sentence-transformers>=3.0
torch>=2.2
numpy>=1.26
transformers>=4.40
Pillow>=10.0
